gutenberg-explorer-api/requirements.txt
fastapi==0.115.6
openapi==2.0.0
python-dotenv==1.0.1
bcrypt==4.2.1
uvicorn==0.34.0
requests==2.32.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
PyJWT==2.10.1
beautifulsoup4==4.12.3